chore: refactor dependency injection and service registration in Main Process
- Removed the Container class and DI_TOKENS for a more straightforward service registration approach.
- Introduced `registerService` for direct service registration in `main.ts`.
- Updated IPC handlers to utilize `getService` for accessing services instead of the DI container.
- Enhanced error handling and logging during service initialization.
- Cleaned up unused imports and improved type safety across IPC controllers.
