+++ src/renderer/components/pages/Favorites.tsx
@@ -106,7 +106,15 @@ export const Favorites = () => {
   const viewType = useSearchStore((state) => state.viewType);
   const filters = useSearchStore((state) => state.filters);
 
-  const { data, fetchNextPage, hasNextPage, isFetchingNextPage, isLoading } =
+  const {
+    data,
+    fetchNextPage,
+    hasNextPage,
+    isFetchingNextPage,
+    isLoading,
+    isError,
+    error,
+  } =
     useInfiniteQuery({
       queryKey: ["posts", "favorites", tags],
       queryFn: async ({ pageParam = 1 }) => {
@@ -159,10 +167,8 @@ export const Favorites = () => {
           // Check if post belongs to tracked artist
           return trackedArtistIds.has(post.artistId);
         });
-      } else {
-        // No tracked artists, show nothing
-        posts = [];
       }
+      // No tracked artists: cannot scope to subscriptions — keep all favorited posts (same as "all")
     } else if (source === "all") {
       // Show all favorites (no filter)
     }
@@ -355,7 +361,14 @@ export const Favorites = () => {
 
       {/* Grid Content */}
       <div className="flex-1 min-h-0">
-        {isLoading && allPosts.length === 0 ? (
+        {isError ? (
+          <div className="flex flex-col justify-center items-center h-full gap-2 px-6 text-center text-destructive">
+            <p className="font-medium">Could not load favorites</p>
+            <p className="text-sm text-muted-foreground max-w-md">
+              {error instanceof Error ? error.message : String(error)}
+            </p>
+          </div>
+        ) : isLoading && allPosts.length === 0 ? (
           <div className="flex justify-center items-center h-full text-muted-foreground">
             <Loader2 className="w-8 h-8 animate-spin" />
           </div>
